Backend/requirements.txt
tensorflow
pandas
numpy
Pillow
google-generativeai
Flask
Flask-Cors
Werkzeug
python-dotenv
Flask-APScheduler
psycopg2-binary
Flask-SQLAlchemy
Flask-Bcrypt
Flask-JWT-Extended
faiss-cpu 
sentence-transformers
tf-keras
openai
bs4
unidecode